Does this change anything?

+++ testgo17.go
@@ -1,6 +1,6 @@
 package testgo17
 
 type SimpleType struct {
-	A int64
-	B uint64	
+	A int32
+	B uint32
 }
